% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_balance.R
\name{addr_balance}
\alias{addr_balance}
\title{The balance of an address.}
\usage{
addr_balance(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the balance of the the address in both, Bitcoin Cash and Satoshi.
}
\description{
Get the balance of a Bitcoin Cash address.
}
\examples{
addr_balance(address)
addr_balance("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
