% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_history.R
\name{addr_history}
\alias{addr_history}
\title{The history of an address.}
\usage{
addr_history(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the general data (hystory) of the the address in both.
}
\description{
Get the general data (hystory) of a Bitcoin Cash address.
}
\examples{
addr_hystory(address)
addr_hystory("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
