% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_totalReceived.R
\name{addr_totalReceived}
\alias{addr_totalReceived}
\title{The total amount of Bitcoin Cash received by an address.}
\usage{
addr_totalReceived(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the total amount of crypto received by the address in both, Bitcoin Cash and Satoshi.
}
\description{
Get the total amount of  Bitcoin Cash received by the address.
}
\examples{
addr_totalReceived(address)
addr_totalReceived("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
