% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_totalSent.R
\name{addr_totalSent}
\alias{addr_totalSent}
\title{The total amount of Bitcoin Cash sent by an address.}
\usage{
addr_totalSent(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the total amount of crypto sent by the address in both, Bitcoin Cash and Satoshi.
}
\description{
Get the total amount of Bitcoin Cash sent by the address.
}
\examples{
addr_totalSent(address)
addr_totalSent("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
