% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_txApperances.R
\name{addr_txApperances}
\alias{addr_txApperances}
\title{The total amount of transaction made by an address.}
\usage{
addr_txApperances(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the total amount of transactions made by the address.
}
\description{
Get the total amount of transactions made by the address.
}
\examples{
addr_txApperances(address)
addr_txApperances("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
