% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_txs.R
\name{addr_txs}
\alias{addr_txs}
\title{All the transactions hashes of a transaction}
\usage{
addr_txs(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with all the transactions made by a Bitcoin Cash address.
}
\description{
Get the all the transactions made by the address.
}
\examples{
addr_txs(address)
addr_txs("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
