% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_unconfirmedTransactions.R
\name{addr_unconfirmedTransactions}
\alias{addr_unconfirmedTransactions}
\title{The total amount of unconfirmed transaction made by an address.}
\usage{
addr_unconfirmedTransactions(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the total amount of unconfirmed transactions made by the address.
}
\description{
Get the total amount of unconfirmed transactions made by the address.
}
\examples{
addr_unconfirmedTransactions(address)
addr_unconfirmedTransactions("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
