% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr_utxo.R
\name{addr_utxo}
\alias{addr_utxo}
\title{The unspent transactions of an address.}
\usage{
addr_utxo(address)
}
\arguments{
\item{string}{input Bitcoin Cash address}
}
\value{
data frame with the unspent transactions of an address.
}
\description{
Get the unspent transactions of an address.
}
\examples{
addr_utxo(address)
addr_utxo("address string here")
}
\author{
Vasile Adrian Bogdan Pop
}
