% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byDate.R
\name{block_byDate}
\alias{block_byDate}
\title{Retrieve the blocks in a specific Date.}
\usage{
block_byDate(date)
}
\arguments{
\item{string}{input Date}
}
\value{
data frame with the blocks of a Date.
}
\description{
Get all the blocks in a specific Date with the format yyyy-MM-dd
}
\examples{
block_byDate("yyyy-MM-dd")
block_byDate(date)
}
\author{
Vasile Adrian Bogdan Pop
}
