% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byDateCount.R
\name{block_byDateCount}
\alias{block_byDateCount}
\title{Retrieve the block count in a specific Date.}
\usage{
block_byDateCount(date)
}
\arguments{
\item{string}{input Date}
}
\value{
data frame with the block count of a Date.
}
\description{
Get the block count in a specific Date with the format yyyy-MM-dd
}
\examples{
block_byDateCount("yyyy-MM-dd")
block_byDateCount(date)
}
\author{
Vasile Adrian Bogdan Pop
}
