% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byHash.R
\name{block_byHash}
\alias{block_byHash}
\title{Retrieve the information contained in a block by his hash.}
\usage{
block_byHash(hash)
}
\arguments{
\item{string}{input hash of a Bitcoin Cash block}
}
\value{
data frame with the information of a specific block.
}
\description{
Get the information of a specific Bitcoin Cash block (without the transactions)
}
\examples{
block_byHash(hash)
block_byHash("block hash here")
}
\author{
Vasile Adrian Bogdan Pop
}
