% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byHashRaw.R
\name{block_byHashRaw}
\alias{block_byHashRaw}
\title{Retrieve the raw informations contained in a block by his hash.}
\usage{
block_byHashRaw(hash)
}
\arguments{
\item{string}{input hash of a Bitcoin Cash block}
}
\value{
data frame with the raw information of a specific block.
}
\description{
Get the raw information of a specific Bitcoin Cash block
}
\examples{
block_byHashRaw(hash)
block_byHashRaw("block hash here")
}
\author{
Vasile Adrian Bogdan Pop
}
