% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byHashTxs.R
\name{block_byHashTxs}
\alias{block_byHashTxs}
\title{Retrieve the transactions hashes contained in a block by his hash.}
\usage{
block_byHashTxs(hash)
}
\arguments{
\item{string}{input hash of a Bitcoin Cash block}
}
\value{
data frame with the transaction hashes of a specific block.
}
\description{
Get the transaction hashes of a specific Bitcoin Cash block
}
\examples{
block_byHashTxs(hash)
block_byHashTxs("block hash here")
}
\author{
Vasile Adrian Bogdan Pop
}
