% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_byHeight.R
\name{block_byHeight}
\alias{block_byHeight}
\title{Retrieve the information contained in a block by his height.}
\usage{
block_byHeight(height)
}
\arguments{
\item{string}{input height of a Bitcoin Cash block}
}
\value{
data frame with the information of a specific block.
}
\description{
Get the informations contained in a specific Bitcoin Cash block
}
\examples{
block_byHeight(height)
block_byHeight("block height here")
}
\author{
Vasile Adrian Bogdan Pop
}
