% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_currency.R
\name{chain_currency}
\alias{chain_currency}
\title{Retrieve the Bitcoin Cash current value.}
\usage{
chain_currency()
}
\value{
data frame with the Bitcoin Cash current value.
}
\description{
Get the Bitcoin Cash current value.
}
\examples{
chain_currency()
}
\author{
Vasile Adrian Bogdan Pop
}
