% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain_stats.R
\name{chain_stats}
\alias{chain_stats}
\title{Retrieve the Bitcoin Cash blockchain stats}
\usage{
chain_stats()
}
\value{
data frame with the Bitcoin Cash blockchain stats
}
\description{
Get the Bitcoin Cash current blockchain stats. To see the description of the values retrieved visit: "https://github.com/Blockchair/Blockchair.Support/blob/master/API_DOCUMENTATION_EN.md#link_chainstats" in Dashboard/stats
}
\examples{
chain_stats()
}
\author{
Vasile Adrian Bogdan Pop
}
