% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txs_byHash.R
\name{txs_byHash}
\alias{txs_byHash}
\title{Retrieve the information about a specific Bitcoin Cash transaction}
\usage{
txs_byHash(hash)
}
\arguments{
\item{string}{hast of the transaction}
}
\value{
data frame with the information of a specific transaction
}
\description{
Get the all the information of a specific transaction. To see the description of the values retrieved visit: "https://github.com/Blockchair/Blockchair.Support/blob/master/API_DOCUMENTATION_EN.md#link_chainstats" in Dashboard/transaction
}
\examples{
txs_byHash()
}
\author{
Vasile Adrian Bogdan Pop
}
