% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txs_byHashIn.R
\name{txs_byHashIn}
\alias{txs_byHashIn}
\title{Retrieve the information about a specific Bitcoin Cash transaction inputs}
\usage{
txs_byHashIn(hash)
}
\arguments{
\item{string}{hast of the transaction inputs}
}
\value{
data frame with the information of a specific transaction inputs
}
\description{
Get the all the information of a specific transaction inputs. To see the description of the values retrieved visit: "https://github.com/Blockchair/Blockchair.Support/blob/master/API_DOCUMENTATION_EN.md#link_chainstats" in Dashboard/transaction
}
\examples{
txs_byHashIn()
}
\author{
Vasile Adrian Bogdan Pop
}
