% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txs_byHashOut.R
\name{txs_byHashOut}
\alias{txs_byHashOut}
\title{Retrieve the information about a specific Bitcoin Cash transaction outputs}
\usage{
txs_byHashOut(hash)
}
\arguments{
\item{string}{hast of the transaction outputs}
}
\value{
data frame with the information of a specific transaction outputs
}
\description{
Get the all the information of a specific transaction outputs. To see the description of the values retrieved visit: "https://github.com/Blockchair/Blockchair.Support/blob/master/API_DOCUMENTATION_EN.md#link_chainstats" in Dashboard/transaction
}
\examples{
txs_byHashOut()
}
\author{
Vasile Adrian Bogdan Pop
}
